79 percent completion
Hitting a Brick that the last time to make it break should return true

Brick.isTouching is not reporting a real BOTTOM collision when the Ball
is near the left of the edge, but still hitting on bottom. Check your
logic by drawing a picture with sample values

    Brick toString method does not match expected result after creation.
Verify that your values are in the right order, and that you are not
off-by-one in computing the pixel position of the bottom right corner.
E.g., remember that the right edge of a brick that is 50 pixels wide and
has a left edge at 10 would be 59, not 60. and that your punctuation
matches Javadoc exactly.